% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_eqM_mean}
\alias{cluster_eqM_mean}
\title{Estimate Population Mean Under Cluster Sampling}
\usage{
cluster_eqM_mean(
  M,
  s2b,
  N = Inf,
  ybari = NULL,
  ybar = NULL,
  n = NULL,
  alpha = 0.05
)
}
\arguments{
\item{M}{a number, the size of each cluster.}

\item{s2b}{a number, the between-sample variance.}

\item{N}{a number, the number of cluster in population.}

\item{ybari}{a vector, the sample mean of each cluster in sample.}

\item{ybar}{a number, the mean of the sample sum of each cluster.}

\item{n}{a number, the number of cluster in sample.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population mean when cluster sampling is used and
the cluster sizes are equal.
}
