% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_prop}
\alias{cluster_prop}
\title{Estimate Population Proportion Under Cluster Sampling}
\usage{
cluster_prop(m, p, N = Inf, ci_method = "ordinary", alpha = 0.05)
}
\arguments{
\item{m}{a number or a vector, the size of each cluster, when the size
is the same, then \code{m} can be simplified as a number.}

\item{p}{a vector, the sample proportion of each cluster in sample.}

\item{N}{a number, the number of cluster in population.}

\item{ci_method}{string, the method to obtain a estimation of the confidence
interval of the estimator, "ordinary" for normal approximation, and "equation"
for equation method.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population proportion when cluster sampling is used.
}
