% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_uneqM_mean}
\alias{cluster_uneqM_mean}
\title{Estimate Population Mean Under Cluster Sampling}
\usage{
cluster_uneqM_mean(
  m,
  yi,
  N,
  M0 = NULL,
  method = "ratio",
  var_method = "mbar",
  ci_method = "ordinary",
  alpha = 0.05
)
}
\arguments{
\item{m}{a vector, the size of each cluster in sample.}

\item{yi}{a vector, the sample sum of each cluster.}

\item{N}{a number, the number of cluster in population.}

\item{M0}{a number, the number of primary units in population.}

\item{method}{string, the estimate method. "srs" means simple estimation,
"ratio" means ratio estimation.}

\item{var_method}{string, the method to obtain a estimation of the variance
of the ratio estimator. If choose "mbar",then will use the mean of \code{m}
directly, else if choose "Mbar", then will use \code{M0}/\code{N}.}

\item{ci_method}{string, the method to obtain a estimation of the confidence
interval of the estimator, "ordinary" for normal approximation, and "equation"
for equation method.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population mean when cluster sampling is used and
the cluster sizes are unequal.
}
