% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_uneqM_total}
\alias{cluster_uneqM_total}
\title{Estimate Population Total Under Cluster Sampling}
\usage{
cluster_uneqM_total(
  m,
  yi,
  N,
  M0 = NULL,
  method = "ratio",
  ci_method = "ordinary",
  alpha = 0.05
)
}
\arguments{
\item{m}{a vector, the size of each cluster in sample.}

\item{yi}{a vector, the sample sum of each cluster.}

\item{N}{a number, the number of cluster in population.}

\item{M0}{a number, the number of secondary units in population.}

\item{method}{string, the estimate method. "srs" means simple estimation,
"ratio" means ratio estimation.}

\item{ci_method}{string, the method to obtain a estimation of the confidence
interval of the estimator, "ordinary" for normal approximation, and "equation"
for equation method.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population total when cluster sampling is used and
the cluster sizes are unequal.
}
