% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_and_reg.R
\name{com_ratio_total}
\alias{com_ratio_total}
\title{Combined Ratio Estimator for Population Total}
\usage{
com_ratio_total(
  Nh = NULL,
  Wh = NULL,
  N = NULL,
  nh,
  ybarh,
  xbarh,
  sy2h,
  sx2h,
  syxh,
  Xbar,
  alpha = 0.05
)
}
\arguments{
\item{Nh}{a vector, the population size of each strata.
If it is be provided, then \code{Wh} and \code{N} can be ignored.}

\item{Wh}{a vector, the weight of each strata.}

\item{N}{a number, the population size.}

\item{nh}{a vector, the sample size of each strata.}

\item{ybarh}{a vector, the strata sample mean of the main variable y.}

\item{xbarh}{a vector, the strata sample mean of the auxiliary variable x.}

\item{sy2h}{a vector, the strata sample variance of the main variable y.}

\item{sx2h}{a vector, the strata sample variance of the auxiliary variable x.}

\item{syxh}{a vector, the strata sample covariance of y and x.}

\item{Xbar}{a number, the historical mean of the auxiliary variable x.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population total using the combined ratio estimator.
}
