% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{conf_interval}
\alias{conf_interval}
\title{Confidence Interval}
\usage{
conf_interval(stat_obs, se_est, alpha = 0.05)
}
\arguments{
\item{stat_obs}{a number, the the observed value of the estimator.}

\item{se_est}{a number, the standard error estimate of the estimator.}

\item{alpha}{a number, the confidence level.}
}
\value{
A list containing:\tabular{ll}{
\code{absolute_error} \tab absolute error limit. \cr
\tab \cr
\code{relative_error} \tab relative error limit. \cr
\tab \cr
\code{ci_left_bound} \tab lower confidence bound. \cr
\tab \cr
\code{ci_right_bound} \tab upper confidence bound. \cr
}
}
\description{
This function is used to estimate the confidence interval for the parameter.
}
\examples{
x <- rnorm(100)
xbar <- mean(x)
se <- sqrt(var(x)/100)
conf_interval(xbar,se,alpha = 0.05)
}
