% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase.R
\name{extract_srs_yxsample}
\alias{extract_srs_yxsample}
\title{Sample Information Extraction}
\usage{
extract_srs_yxsample(yxsample, yloc)
}
\arguments{
\item{yxsample}{matrix or dataframe, which have two columns. One for y,
another for x.}

\item{yloc}{a number, the location of y column in the dataframe.}
}
\description{
This function is an auxiliary function which can extract the sample mean, sample variance and sample covariance of the
main variable y and the auxiliary variable x.
}
