% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequal.R
\name{PiPS_HT_mean}
\alias{PiPS_HT_mean}
\title{Horvitz-Thompson Estimator for Population Total}
\usage{
PiPS_HT_mean(y, PI, PIM, N, alpha = 0.05)
}
\arguments{
\item{y}{a vector, the sample data.}

\item{PI}{a vector, the inclusion probability.}

\item{PIM}{a matrix, the joint inclusion probability.}

\item{N}{a number, the size of population.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population total with Horvitz-Thompson estimator
when PPS sampling is used.
}
