% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequal.R
\name{PiPS_HT_total}
\alias{PiPS_HT_total}
\title{Horvitz-Thompson Estimator for Population Total}
\usage{
PiPS_HT_total(y, PI, PIM, alpha = 0.05)
}
\arguments{
\item{y}{a vector, the sample data.}

\item{PI}{a vector, the inclusion probability probability.}

\item{PIM}{a matrix, the joint inclusion probability.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population total with Horvitz-Thompson estimator
when PPS sampling is used.
}
