% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequal.R
\name{PPS_HH_mean}
\alias{PPS_HH_mean}
\title{Hansen-Hurwitz Estimator for Population Mean}
\usage{
PPS_HH_mean(y, z, N, alpha = 0.05)
}
\arguments{
\item{y}{a vector, the sample data.}

\item{z}{a vector, the probability that the corresponding sample is selected
in a single sampling.}

\item{N}{a number, the size of population.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population mean with Hansen-Hurwitz estimator
when PPS sampling is used.
}
