% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_and_reg.R
\name{ratio}
\alias{ratio}
\title{Estimate Ratio}
\usage{
ratio(
  ybar,
  xbar,
  sy2,
  sx2,
  syx,
  n,
  N = Inf,
  var_method = "xbar",
  ci_method = "ordinary",
  Xbar = NULL,
  alpha = 0.05
)
}
\arguments{
\item{ybar}{a number, the sample mean of the main variable y.}

\item{xbar}{a number, the sample mean of the auxiliary variable x.}

\item{sy2}{a number, the sample variance of the main variable y.}

\item{sx2}{a number, the sample variance of the auxiliary variable x.}

\item{syx}{a number, the sample covariance of y and x.}

\item{n}{a number, the sample size.}

\item{N}{a number, the population size.}

\item{var_method}{string, the method to obtain a estimation of the variance
of the estimator. If choose "xbar",then will use \code{xbar} directly, else if
choose "Xbar", then \code{Xbar} needs be given.}

\item{ci_method}{string, the method to obtain a estimation of the confidence
interval of the estimator, "ordinary" for normal approximation, and "equation"
for equation method.}

\item{Xbar}{a number, the historical mean of the auxiliary variable x.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the ratio between the main variable y and the auxiliary variable x.
}
