% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_and_reg.R
\name{ratio_mean}
\alias{ratio_mean}
\title{Estimate Population Mean with Ratio Estimation}
\usage{
ratio_mean(
  ybar,
  xbar,
  sy2,
  sx2,
  syx,
  n,
  Xbar,
  N = Inf,
  ci_method = "ordinary",
  alpha = 0.05
)
}
\arguments{
\item{ybar}{a number, the sample mean of the main variable y.}

\item{xbar}{a number, the sample mean of the auxiliary variable x.}

\item{sy2}{a number, the sample variance of the main variable y.}

\item{sx2}{a number, the sample variance of the auxiliary variable x.}

\item{syx}{a number, the sample covariance of y and x.}

\item{n}{a number, the sample size.}

\item{Xbar}{a number, the historical mean of the auxiliary variable x.}

\item{N}{a number, the population size.}

\item{ci_method}{string, the method to obtain a estimation of the confidence
interval of the estimator, "ordinary" for normal approximation, and "equation"
for equation method.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function uses ratio estimation to estimate the population mean.
}
