% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_and_reg.R
\name{ratio_sample_extract}
\alias{ratio_sample_extract}
\title{Sample Information Extraction}
\usage{
ratio_sample_extract(y_sample, x_sample)
}
\arguments{
\item{y_sample}{a vector, the sample of the main variable y.}

\item{x_sample}{a vector, the sample of the auxiliary variable x.}
}
\description{
This function is an auxiliary function for ratio estimation and regression estimation
and can extract the sample mean, sample variance and sample covariance of the
main variable y and the auxiliary variable x.
}
