% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_and_reg.R
\name{reg_total}
\alias{reg_total}
\title{Estimate Population Total with Regression Estimation}
\usage{
reg_total(
  ybar,
  xbar,
  sy2,
  sx2,
  syx,
  n,
  Xbar,
  N,
  beta_method = "reg_coef",
  beta0 = NULL,
  alpha = 0.05
)
}
\arguments{
\item{ybar}{a number, the sample mean of the main variable y.}

\item{xbar}{a number, the sample mean of the auxiliary variable x.}

\item{sy2}{a number, the sample variance of the main variable y.}

\item{sx2}{a number, the sample variance of the auxiliary variable x.}

\item{syx}{a number, the sample covariance of y and x.}

\item{n}{a number, the sample size.}

\item{Xbar}{a number, the historical mean of the auxiliary variable x.}

\item{N}{a number, the population size.}

\item{beta_method}{string, coefficient of auxiliary variables x in regression estimation.
If choose "reg_coef", then the sample regression coefficient will be used.
If choose "const", then \code{beta0} need be given and be used.}

\item{beta0}{a number.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function uses regression estimation to estimate the population total.
}
