% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unequal.R
\name{RHC_mean}
\alias{RHC_mean}
\title{Estimate Population Mean Under Rao-Hartley-Cochran sampling}
\usage{
RHC_mean(ysample, yprob, group_prob_sum, group_N, alpha = 0.05)
}
\arguments{
\item{ysample}{a vector, the sample data.}

\item{yprob}{a vector, the probability that the corresponding sample is selected
into sample.}

\item{group_prob_sum}{a vector, the sum of probability in a group where the
sample is selected.}

\item{group_N}{a vector, the size of the group where the sample is selected.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population mean when RHC sampling is used.
}
