% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_and_reg.R
\name{sep_reg_total}
\alias{sep_reg_total}
\title{Separate Regression Estimator for Population Total}
\usage{
sep_reg_total(
  Nh = NULL,
  Wh = NULL,
  N = NULL,
  nh,
  ybarh,
  xbarh,
  sy2h,
  sx2h,
  syxh,
  Xbarh,
  beta_method = "reg_coef",
  alpha = 0.05,
  betah0 = NULL,
  stra_est = FALSE
)
}
\arguments{
\item{Nh}{a vector, the population size of each strata.
If it is be provided, then \code{Wh} and \code{N} can be ignored.}

\item{Wh}{a vector, the weight of each strata.}

\item{N}{a number, the population size.}

\item{nh}{a vector, the sample size of each strata.}

\item{ybarh}{a vector, the strata sample mean of the main variable y.}

\item{xbarh}{a vector, the strata sample mean of the auxiliary variable x.}

\item{sy2h}{a vector, the strata sample variance of the main variable y.}

\item{sx2h}{a vector, the strata sample variance of the auxiliary variable x.}

\item{syxh}{a vector, the strata sample covariance of y and x.}

\item{Xbarh}{a vector, the historical strata mean of the auxiliary variable x.}

\item{beta_method}{string, coefficient of auxiliary variables x in regression estimation.
If choose "reg_coef", then the strata sample regression coefficient will be used.
If choose "const", then \code{beta0} need be given and be used.}

\item{alpha}{a number, the confidence level.}

\item{betah0}{a number or a vector. If the coefficient for each stata is the
same, then can just input a number.}

\item{stra_est}{logical, whether to calculate and output the estimated results
of each strata.}
}
\description{
This function estimates the population total using the separate regression estimator.
}
