% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs_mean}
\alias{srs_mean}
\title{Estimate Population Mean}
\usage{
srs_mean(sample_data, N = Inf, alpha = 0.05)
}
\arguments{
\item{sample_data}{a vector, the sample data.}

\item{N}{a number, the size of population, If it is unknown or very large, it can be omitted.
By default, infinity is used for calculation.}

\item{alpha}{a number, the confidence level.}
}
\value{
A list containing:\tabular{ll}{
\code{ybar_result} \tab a list containing estimation of the population mean, estimation of the variance of the estimator and the estimation of the standard error of the estimator. \cr
\tab \cr
\code{ci_est} \tab a list contain absolute/relative error limit and the estimation of lower/upper confidence bound of the estimator.\cr
\tab \cr
}
}
\description{
This function is used to estimate the population mean in simple random sampling.
}
