% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs_num}
\alias{srs_num}
\title{Estimate the Number of Units with a Certain Type}
\usage{
srs_num(event_num, n, N, alpha = 0.05)
}
\arguments{
\item{event_num}{a number, the number of cells in the sample with a certain characteristic.}

\item{n}{a number, the sample size}

\item{N}{a number, the size of population.}

\item{alpha}{a number, the confidence level.}
}
\value{
A list containing:\tabular{ll}{
\code{a_result} \tab a list containing estimation of the number of units in the population that have a certain characteristic, estimation of the variance of the estimator and the estimation of the standard error of the estimator. \cr
\tab \cr
\code{ci_est} \tab a list contain absolute/relative error limit and the estimation of lower/upper confidence bound of the estimator.\cr
\tab \cr
}
}
\description{
This function is used to estimate The number of units in the population that have a certain characteristic.
}
