% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs_prop}
\alias{srs_prop}
\title{Estimate Population Proportion}
\usage{
srs_prop(event_num, n, N = Inf, alpha = 0.05)
}
\arguments{
\item{event_num}{a number, the number of units in the sample with a certain characteristic.}

\item{n}{a number, the sample size}

\item{N}{a number, the size of population, If it is unknown or very large, it can be omitted.
By default, infinity is used for calculation.}

\item{alpha}{a number, the confidence level.}
}
\value{
A list containing:\tabular{ll}{
\code{p_result} \tab a list containing estimation of the population proportion, estimation of the variance of the estimator and the estimation of the standard error of the estimator. \cr
\tab \cr
\code{ci_est} \tab a list contain absolute/relative error limit and the estimation of lower/upper confidence bound of the estimator.\cr
\tab \cr
}
}
\description{
This function is used to estimate The proportion of units in the population that have a certain characteristic.
}
