% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srs_size_for_mean}
\alias{srs_size_for_mean}
\title{Sample Size for Mean}
\usage{
srs_size_for_mean(
  var_his,
  mean_his = NULL,
  N = Inf,
  method,
  bound,
  alpha = 0.05
)
}
\arguments{
\item{var_his}{a number, historical estimates of population variation.}

\item{mean_his}{a number, historical estimates of population mean. It need be
specified if the precision requirement is a relative error limit or an upper
coefficient of variation limit.}

\item{N}{a number, the size of population, If it is unknown or very large, it can be omitted.
By default, infinity is used for calculation.}

\item{method}{string, the form of precision requirement.
For upper limit of variance, use "V", for absolute error limit, use "d",
for relative error limit, use "r", for upper limit of variation coefficient,
use "CV".}

\item{bound}{a number, the upper limit value of the precision requirement.}

\item{alpha}{a number, the confidence level.}
}
\value{
a list contain the form of precision requirement, uncorrected sample
size and corrected sample size.
}
\description{
This function estimates the sample size required for a given mean estimator precision.
}
\note{
If the total estimator precision is given, convert it to the mean estimator
precision then use this function. This function will convert all forms of upper precision to
upper limit of variance in calculation.
}
