% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop_and_poststra.R
\name{srs_subpop_total1}
\alias{srs_subpop_total1}
\title{Estimating the Subpopulation Total}
\usage{
srs_subpop_total1(y_sub, n, N, alpha = 0.05)
}
\arguments{
\item{y_sub}{a vector, a sample of a subpopulation.}

\item{n}{a number, the sample size of the entire population.}

\item{N}{a number, the size of the entire population.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function is used to estimate the subpopulation total.
}
\note{
This function is for the data that have origin data.
}
