% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop_and_poststra.R
\name{srs_subpop_total2}
\alias{srs_subpop_total2}
\title{Estimating the Subpopulation Total}
\usage{
srs_subpop_total2(ybar_sub, s2_sub, n_sub, n, N, alpha = 0.05)
}
\arguments{
\item{ybar_sub}{a number, the sample mean of a subpopulation.}

\item{s2_sub}{a number, the sample variance of a subpopulation.}

\item{n_sub}{a number, the sample size of a subpopulation.}

\item{n}{a number, the sample size of the entire population.}

\item{N}{a number, the size of the entire population.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function is used to estimate the subpopulation total.
}
\note{
This function is for the data that only have some information from the
origin data.
}
