% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra.R
\name{stra_allocation}
\alias{stra_allocation}
\title{Sample Size Allocation}
\usage{
stra_allocation(
  Nh,
  S2h = NULL,
  Ph = NULL,
  ch = NULL,
  n,
  allocation,
  type = "mean"
)
}
\arguments{
\item{Nh}{a vector, the size of each population strata.}

\item{S2h}{a vector, the variance of each population strata, which can be
estimated by historical data. It can be ignored when \code{type} is "prop".}

\item{Ph}{a vector, the proportion of each population strata, which can be
estimated by historical data. It can be ignored when \code{type} is "mean".}

\item{ch}{a vector, the unit cost of each strata. It can be just the cost weights instead of specific values.}

\item{n}{a number, the sample size one wishes.}

\item{allocation}{string, the allocation method. "Prop" denotes proportional
allocation, "Opt" denotes optimal allocation, "Neyman" demotes Neyman allocation.}

\item{type}{string, the type of the parameter to be estimated,
"mean" for population mean and "prop" for population proportion.}
}
\description{
This function is used to calculate the sample size of each strata under the given sample allocation method.
}
