% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra.R
\name{stra_sample_generate}
\alias{stra_sample_generate}
\title{Sampling simulation}
\usage{
stra_sample_generate(pop_data, n, wh = NULL)
}
\arguments{
\item{pop_data}{a list, each element denotes a strata population.}

\item{n}{a number, the sample size one wishes.}

\item{wh}{a vector, the allocation proportion of each stratum sample. It will be normalized.
If it is not given, then the function will use proportional allocation.}
}
\description{
A function for simulation calculations that draws samples from the population.
}
\note{
Since there is rounding error, the final sample size may not be \code{n}.
}
