% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra.R
\name{stra_size}
\alias{stra_size}
\title{Estimate Sample size}
\usage{
stra_size(
  Nh,
  Wh = NULL,
  varh_his = NULL,
  mean_his = NULL,
  Ph_his,
  ch = NULL,
  P_his = NULL,
  method,
  bound,
  alpha = 0.05,
  allocation,
  type = "mean"
)
}
\arguments{
\item{Nh}{a vector, the size of each population strata.}

\item{Wh}{a vector, the weight of each population strata. It can be ignored
since it is for the case N is infinity where \code{Nh} cannot be specified.}

\item{varh_his}{a vector, historical estimates of each strata population variance.}

\item{mean_his}{a vector, historical estimates of each strata population mean.}

\item{Ph_his}{a vector, historical estimates of each strata population proportion.}

\item{ch}{a vector, the unit cost of each strata. It can be just the cost weights instead of specific values.}

\item{P_his}{a number, a vector, historical estimates of population prportion.}

\item{method}{string, the form of precision requirement.
For upper limit of variance, use "V", for absolute error limit, use "d",
for relative error limit, use "r", for upper limit of variation coefficient,
use "CV".}

\item{bound}{a number, the upper limit value of the precision requirement.}

\item{alpha}{a number, the confidence level.}

\item{allocation}{string, the allocation method. "Prop" denotes proportional
allocation, "Opt" denotes optimal allocation, "Neyman" demotes Neyman allocation.}

\item{type}{string, the type of the parameter to be estimated,
"mean" for population mean and "prop" for population proportion.}
}
\value{
a list contain the form of precision requirement, uncorrected sample
size and corrected sample size.
}
\description{
#' This function estimates the sample size required for a given mean/proportion estimator precision.
}
