% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra.R
\name{stra_srs_mean}
\alias{stra_srs_mean}
\title{Estimate Population Mean}
\usage{
stra_srs_mean(
  Nh = NULL,
  Wh = NULL,
  N = NULL,
  nh,
  ybarh,
  s2h,
  alpha = 0.05,
  stra_est = FALSE
)
}
\arguments{
\item{Nh}{a vector, the population size of each strata.
If it is be provided, then \code{Wh} and \code{N} can be ignored.}

\item{Wh}{a vector, the weight of each strata.}

\item{N}{a number, the population size.}

\item{nh}{a vector, the sample size of each strata.}

\item{ybarh}{a vector, the sample mean of each strata.}

\item{s2h}{a vector, the sample variance of each strata.}

\item{alpha}{a number, the confidence level.}

\item{stra_est}{logical, whether to calculate and output the estimated results
of each strata.}
}
\description{
This function is used to estimate the population mean in stratified simple random sampling.
}
