% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra.R
\name{stra_srs_prop}
\alias{stra_srs_prop}
\title{Estimate Population Proportion}
\usage{
stra_srs_prop(
  Nh = NULL,
  Wh = NULL,
  N = NULL,
  nh,
  ph,
  alpha = 0.05,
  stra_est = FALSE
)
}
\arguments{
\item{Nh}{a vector, the population size of each strata.
If it is be provided, then \code{Wh} and \code{N} can be ignored.}

\item{Wh}{a vector, the weight of each strata.}

\item{N}{a number, the population size.}

\item{nh}{a vector, the sample size of each strata.}

\item{ph}{a vector, the sample proportion of each strata.}

\item{alpha}{a number, the confidence level.}

\item{stra_est}{logical, whether to calculate and output the estimated results
of each strata.}
}
\description{
This function is used to estimate The proportion of units in the population
that have a certain characteristic in stratified simple random sampling.
}
