% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage.R
\name{two_stage_eqM_prop}
\alias{two_stage_eqM_prop}
\title{Estimate Population Proportion Under Two-Stage Sampling}
\usage{
two_stage_eqM_prop(M, m, pj, N = Inf, alpha = 0.05)
}
\arguments{
\item{M}{a number, the number of secondary units in each primary unit in population.}

\item{m}{a number, the number of secondary units in each primary unit in sample.}

\item{N}{a number, the number of primary units in population.}

\item{alpha}{a number, the confidence level.}

\item{s2i}{a vector, the sample proportion of each primary unit in sample.}
}
\description{
This function estimates the population proportion when two-stage sampling is used and
the cluster sizes are equal.
}
