% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage.R
\name{two_stage_uneqM_total}
\alias{two_stage_uneqM_total}
\title{Estimate Population Total Under Two-Stage Sampling}
\usage{
two_stage_uneqM_total(M0, M2i, mi, ybari, s2i, N, alpha = 0.05)
}
\arguments{
\item{M0}{a number, the number of secondary units in population.}

\item{M2i}{a vector, the number of secondary units in each primary unit that is drawn into sample.}

\item{mi}{a vector, the number of secondary units in each primary unit in sample.}

\item{ybari}{a vector, the sample mean of each primary unit in sample.}

\item{s2i}{a vector, the sample variance of each primary unit in sample.}

\item{N}{a number, the number of primary units in population.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population total when two-stage sampling is used and
the cluster sizes are unequal.
}
