% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase.R
\name{twophase_ratio_total}
\alias{twophase_ratio_total}
\title{Estimate Population Total Under Double Sampling with Ratio Estimation}
\usage{
twophase_ratio_total(n1, n2, xbar1, ybar, xbar, sx2, sy2, syx, N, alpha = 0.05)
}
\arguments{
\item{n1}{a vector, the sample size of the first phase sample.}

\item{n2}{a vector, the sample size of the second phase sample.}

\item{xbar1}{a number, the sample mean of the auxiliary variable x in the
first phase sample.}

\item{ybar}{a number, the sample mean of the main variable y in the second
phase sample.}

\item{xbar}{a number, the sample mean of the auxiliary variable x in the
second phase sample}

\item{sx2}{a number, the sample variance of the main variable y in the second
phase sample.}

\item{sy2}{a number, the sample variance of the auxiliary variable x in the
second phase sample.}

\item{syx}{a number, the sample covariance of y and x in the second phase sample.}

\item{N}{a number, the population size.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population total when double sampling that the first
sample is only used for the auxiliary variable x and the second sample is used
for the main variable y.
}
