% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_phase.R
\name{twophase_stra_mean}
\alias{twophase_stra_mean}
\title{Estimate Population Mean Under Double Sampling with Stratification}
\usage{
twophase_stra_mean(nh1, nh2, ybarh, s2h, N = Inf, alpha = 0.05)
}
\arguments{
\item{nh1}{a vector, the strata sample size of the first phase sample.}

\item{nh2}{a vector, the strata sample size of the second phase sample.}

\item{ybarh}{a vector, the strata sample mean of the second phase sample.}

\item{s2h}{a vector, the strata sample variance of the second phase sample.}

\item{N}{a number, the population size.}

\item{alpha}{a number, the confidence level.}
}
\description{
This function estimates the population mean when double sampling that the first
sampling method is simple random sample sample and the second sampling method
is stratified sample is used.
}
